pyproject.toml
[project]
name = "auto-anki-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
